+++ app/Http/Controllers/VillageheadController.php
@@ -9,6 +9,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Request as FReq;
 use App\Models\District;
 use App\Models\Villagehead;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 // use Illuminate\Http\Request;
 
 use Inertia\Inertia;
@@ -25,6 +27,10 @@ class VillageheadController extends Controller
     public function index()
     {
         //
+        $editDetails = true;
+        if (!Gate::allows('edit-data',Auth::user()->role)){
+            $editDetails = false;
+        }
         $villageheads = villagehead::query()
             ->when(FReq::input('search'), function ($query, $search) {
                 $query->where('incumbent', 'like', "%{$search}%");
@@ -41,7 +47,7 @@ class VillageheadController extends Controller
                 'slug' => $villagehead->slug,
             ]);
         $filters = FReq::only(['search']);
-        return Inertia::render("Backend/VillageHead/Index", compact('villageheads', 'filters'));
+        return Inertia::render("Townboard/VillageHead/Index", compact('villageheads', 'filters','editDetails'));
     }
 
     /**
@@ -52,7 +58,11 @@ class VillageheadController extends Controller
     public function create()
     {
         //
-        return Inertia::render("Backend/VillageHead/Create");
+
+        if (!Gate::allows('edit-data',Auth::user()->role)){
+            abort(403);
+        }
+        return Inertia::render("Townboard/VillageHead/Create");
     }
 
     /**
@@ -63,7 +73,9 @@ class VillageheadController extends Controller
      */
     public function store(Request $request)
     {
-
+         if (!Gate::allows('edit-data',Auth::user()->role)){
+            abort(403);
+        }
         $validated = $request->validate([
             'district' => 'required',
             'province' => 'required',
@@ -112,7 +124,7 @@ class VillageheadController extends Controller
     public function show($slug)
     {
         $villagehead = Villagehead::where('slug', $slug)->first();
-        return Inertia::render("Backend/VillageHead/Show", compact('villagehead'));
+        return Inertia::render("Townboard/VillageHead/Show", compact('villagehead'));
 
     }
 
@@ -125,8 +137,10 @@ class VillageheadController extends Controller
     public function edit(Villagehead $villagehead)
     {
         //
-
-        return Inertia::render('Backend/VillageHead/Edit', compact('villagehead'));
+         if (!Gate::allows('edit-data',Auth::user()->role)){
+            abort(403);
+        }
+        return Inertia::render('Townboard/VillageHead/Edit', compact('villagehead'));
     }
 
     /**
@@ -139,6 +153,9 @@ class VillageheadController extends Controller
     public function update(Request $request, Villagehead $villagehead)
     {
         //
+         if (!Gate::allows('edit-data',Auth::user()->role)){
+            abort(403);
+        }
         $validated = $request->validate([
             'district' => 'required',
             'province' => 'required',
@@ -189,6 +206,9 @@ class VillageheadController extends Controller
     public function destroy(Villagehead $villagehead)
     {
         //
+        if (!Gate::allows('edit-data',Auth::user()->role)){
+            abort(403);
+        }
         $villagehead->delete();
         return back()->with('message', 'villagehead Deleted Successfull');
     }
